[TASK] Merge `AbstractEventMailForm` into `GeneralEventMailForm`
Fixes #3952
