--- Tests/Functional/BackEnd/Fixtures/TestingEventMailForm.php
@@ -1,14 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace OliverKlee\Seminars\Tests\Functional\BackEnd\Fixtures;
-
-use OliverKlee\Seminars\BackEnd\AbstractEventMailForm;
-
-/**
- * This class represents a testing implementation of the AbstractEventMailForm class.
- */
-final class TestingEventMailForm extends AbstractEventMailForm
-{
-}
