+++ Tests/Functional/BackEnd/GeneralEventMailFormTest.php
@@ -14,7 +14,6 @@ use TYPO3\CMS\Core\Utility\GeneralUtility;
 use TYPO3\TestingFramework\Core\Functional\FunctionalTestCase;
 
 /**
- * @covers \OliverKlee\Seminars\BackEnd\AbstractEventMailForm
  * @covers \OliverKlee\Seminars\BackEnd\GeneralEventMailForm
  */
 final class GeneralEventMailFormTest extends FunctionalTestCase
@@ -28,6 +27,13 @@ final class GeneralEventMailFormTest extends FunctionalTestCase
         'typo3conf/ext/feuserextrafields',
         'typo3conf/ext/oelib',
         'typo3conf/ext/seminars',
+    ];
+
+    protected array $configurationToUseInTestInstance = [
+        'MAIL' => [
+            'defaultMailFromAddress' => 'system-foo@example.com',
+            'defaultMailFromName' => 'Mr. Default',
+        ],
     ];
 
     protected function setUp(): void
@@ -48,6 +54,207 @@ final class GeneralEventMailFormTest extends FunctionalTestCase
         unset($GLOBALS['TYPO3_CONF_VARS']['EXTCONF']['seminars']['backEndModule']);
 
         parent::tearDown();
+    }
+
+    /**
+     * @test
+     */
+    public function sendEmailToAttendeesForTwoRegistrationsSendsTwoEmails(): void
+    {
+        $this->importDataSet(__DIR__ . '/Fixtures/Records.xml');
+
+        $this->email->expects(self::exactly(2))->method('send');
+        $this->addMockedInstance(MailMessage::class, $this->email);
+        $this->addMockedInstance(MailMessage::class, $this->email);
+
+        $subject = new GeneralEventMailForm(2);
+
+        $subject->setPostData(
+            [
+                'action' => 'sendEmail',
+                'isSubmitted' => '1',
+                'subject' => 'foo',
+                'messageBody' => 'some message body',
+            ]
+        );
+        $subject->sendEmailToAttendees();
+    }
+
+    /**
+     * @test
+     */
+    public function sendEmailToAttendeesSendsEmailWithNameOfRegisteredUserInSalutationMarker(): void
+    {
+        $this->importDataSet(__DIR__ . '/Fixtures/Records.xml');
+
+        $this->email->expects(self::once())->method('send');
+        $this->addMockedInstance(MailMessage::class, $this->email);
+
+        $subject = new GeneralEventMailForm(1);
+
+        $messageBody = '%salutation';
+        $subject->setPostData(
+            [
+                'action' => 'sendEmail',
+                'isSubmitted' => '1',
+                'subject' => 'foo',
+                'messageBody' => $messageBody,
+            ]
+        );
+        $subject->sendEmailToAttendees();
+
+        self::assertStringContainsString('Joe Johnson', $this->email->getTextBody());
+    }
+
+    /**
+     * @test
+     */
+    public function sendEmailToAttendeesUsesTypo3DefaultFromAddressAsSender(): void
+    {
+        $this->importDataSet(__DIR__ . '/Fixtures/Records.xml');
+
+        $subject = new GeneralEventMailForm(2);
+
+        $subject->setPostData(
+            [
+                'action' => 'sendEmail',
+                'isSubmitted' => '1',
+                'subject' => 'foo',
+                'messageBody' => 'Hello!',
+            ]
+        );
+
+        $this->email->expects(self::exactly(2))->method('send');
+        $this->addMockedInstance(MailMessage::class, $this->email);
+        $this->addMockedInstance(MailMessage::class, $this->email);
+        $subject->sendEmailToAttendees();
+
+        self::assertArrayHasKey('system-foo@example.com', $this->getFromOfEmail($this->email));
+    }
+
+    /**
+     * @test
+     */
+    public function sendEmailToAttendeesUsesFirstOrganizerAsSender(): void
+    {
+        $GLOBALS['TYPO3_CONF_VARS']['MAIL'] = [];
+
+        $this->importDataSet(__DIR__ . '/Fixtures/Records.xml');
+
+        $subject = new GeneralEventMailForm(2);
+
+        $subject->setPostData(
+            [
+                'action' => 'sendEmail',
+                'isSubmitted' => '1',
+                'subject' => 'foo',
+                'messageBody' => 'Hello!',
+            ]
+        );
+        $this->email->expects(self::exactly(2))->method('send');
+        $this->addMockedInstance(MailMessage::class, $this->email);
+        $this->addMockedInstance(MailMessage::class, $this->email);
+        $subject->sendEmailToAttendees();
+
+        self::assertArrayHasKey('oliver@example.com', $this->getFromOfEmail($this->email));
+    }
+
+    /**
+     * @test
+     */
+    public function sendEmailToAttendeesEmailUsesFirstOrganizerAsReplyTo(): void
+    {
+        $this->importDataSet(__DIR__ . '/Fixtures/Records.xml');
+
+        $subject = new GeneralEventMailForm(2);
+
+        $subject->setPostData(
+            [
+                'action' => 'sendEmail',
+                'isSubmitted' => '1',
+                'subject' => 'foo',
+                'messageBody' => 'Hello!',
+            ]
+        );
+        $this->email->expects(self::exactly(2))->method('send');
+        $this->addMockedInstance(MailMessage::class, $this->email);
+        $this->addMockedInstance(MailMessage::class, $this->email);
+        $subject->sendEmailToAttendees();
+
+        self::assertArrayHasKey('oliver@example.com', $this->getReplyToOfEmail($this->email));
+    }
+
+    /**
+     * @test
+     */
+    public function sendEmailToAttendeesEmailAppendsFirstOrganizerFooterToMessageBody(): void
+    {
+        $this->importDataSet(__DIR__ . '/Fixtures/Records.xml');
+
+        $this->email->expects(self::exactly(2))->method('send');
+        $this->addMockedInstance(MailMessage::class, $this->email);
+        $this->addMockedInstance(MailMessage::class, $this->email);
+
+        $subject = new GeneralEventMailForm(2);
+
+        $subject->setPostData(
+            [
+                'action' => 'sendEmail',
+                'isSubmitted' => '1',
+                'subject' => 'foo',
+                'messageBody' => 'Hello!',
+            ]
+        );
+        $subject->sendEmailToAttendees();
+
+        self::assertStringContainsString("\n-- \nThe one and only", $this->email->getTextBody());
+    }
+
+    /**
+     * @test
+     */
+    public function sendEmailToAttendeesUsesProvidedEmailSubject(): void
+    {
+        $this->importDataSet(__DIR__ . '/Fixtures/Records.xml');
+
+        $this->email->expects(self::once())->method('send');
+        $this->addMockedInstance(MailMessage::class, $this->email);
+
+        $emailSubject = 'Thank you for your registration.';
+        $subject = new GeneralEventMailForm(1);
+        $subject->setPostData(
+            [
+                'action' => 'sendEmail',
+                'isSubmitted' => '1',
+                'subject' => $emailSubject,
+                'messageBody' => 'Hello!',
+            ]
+        );
+        $subject->sendEmailToAttendees();
+
+        self::assertSame($emailSubject, $this->email->getSubject());
+    }
+
+    /**
+     * @test
+     */
+    public function sendEmailToAttendeesNotSendsEmailToUserWithoutEmailAddress(): void
+    {
+        $this->importDataSet(__DIR__ . '/Fixtures/Records.xml');
+
+        $this->email->expects(self::never())->method('send');
+        $this->addMockedInstance(MailMessage::class, $this->email);
+
+        $subject = new GeneralEventMailForm(4);
+        $subject->setPostData(
+            [
+                'action' => 'sendEmail',
+                'isSubmitted' => '1',
+                'subject' => 'Hello!',
+                'messageBody' => 'Hello!',
+            ]
+        );
+        $subject->sendEmailToAttendees();
     }
 
     /**
